+++ src/imio/smartweb/core/browser/faceted/views.py
@@ -3,6 +3,7 @@
 from eea.facetednavigation.browser.app.view import FacetedContainerView
 from eea.facetednavigation.layout.interfaces import IFacetedLayout
 from imio.smartweb.core.interfaces import IViewWithoutLeadImage
+from imio.smartweb.core.utils import get_scale_url
 from plone.app.contenttypes.browser.folder import FolderView
 from zope.component import queryAdapter
 from zope.interface import implementer
@@ -29,13 +30,14 @@ class FacetedView(FolderView):
         return item.portal_type == "imio.smartweb.SectionVideo"
 
     def get_image_url(self, item):
+        scale_url = ""
         if item.portal_type == "imio.smartweb.SectionGallery":
             images = item.getObject().listFolderContents()
             if not images:
                 return ""
-            url = images[0].absolute_url()
-            return f"{url}/@@images/image"
+            scale_url = get_scale_url(item, self.request, "image", "")
+            return scale_url
         if not item.has_leadimage:
             return ""
         url = item.getURL()
-        return f"{url}/@@images/image"
+        return scale_url