+++ src/imio/smartweb/core/contents/sections/links/view.py
@@ -2,6 +2,7 @@
 
 from imio.smartweb.core.contents.sections.views import CarouselOrTableSectionView
 from imio.smartweb.core.utils import batch_results
+from imio.smartweb.core.utils import get_scale_url
 
 
 class LinksView(CarouselOrTableSectionView):
@@ -18,6 +19,7 @@ class LinksView(CarouselOrTableSectionView):
                 has_icon = True
             elif getattr(item.aq_base, "image", None):
                 has_image = True
+            scale_url = get_scale_url(item, self.request, "image", image_scale)
             results.append(
                 {
                     "title": item.title,
@@ -25,7 +27,7 @@ class LinksView(CarouselOrTableSectionView):
                     "url": url,
                     "icon": item.svg_icon,
                     "has_icon": has_icon,
-                    "image": f"{url}/@@images/image/{image_scale}",
+                    "image": scale_url,
                     "has_image": has_image,
                     "open_in_new_tab": item.open_in_new_tab,
                 }